build: Update dependencies
Breaking changes dealt with:
* fmt stopped implicitly converting scoped enums to ints, so static
  casts were added to fix that.
* ftxui changed ftxui::paragraph to return a single element instead of a
  list of them, so our hack of wrapping that list in a hflow is no
  longer needed.

+++ browser/tui.cpp
@@ -29,7 +29,7 @@ int main(int argc, char **argv) {
 
     browser::Engine engine;
     if (auto err = engine.navigate(*uri); err != protocol::Error::Ok) {
-        spdlog::error("Got error {} from {}", err, uri->uri);
+        spdlog::error("Got error {} from {}", static_cast<int>(err), uri->uri);
         std::exit(1);
     }
